refactor: 당첨번호와 보너스번호를 갖는 WinningLotto 클래스 생성 & Ranks 일급 컬렉션 생성

+++ src/main/java/lotto/domain/Amount.java
@@ -1,13 +1,13 @@
 package lotto.domain;
 
 public class Amount {
-    private final int value;
+    private final long value;
 
-    public Amount(int value) {
+    public Amount(long value) {
         this.value = value;
     }
 
-    public int divide(int divisor) {
+    public long divide(long divisor) {
         return value / divisor;
     }
 }
